Refactor out NoteFragment test code to NoteFragmentFeatureUtilTest.
